backend/requirements.txt
Flask==3.0.3
fastapi==0.115.0
uvicorn==0.32.0
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
alembic==1.13.1